pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradebot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "requests==2.32.3",
  "PyJWT==2.8.0",
  "cryptography==42.0.5",
  "python-dotenv==1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]